refactor factory upgrades to a service
